pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel",
    "numpy>=1.19.0"
]
build-backend = "setuptools.build_meta"

[project]
name = "ezvtuber-rt"
version = "0.1.0"
description = "Real-time virtual tuber animation library with TensorRT and ONNX support"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "zpeng11"}
]
keywords = [
    "vtuber",
    "real-time",
    "animation",
    "tensorrt",
    "onnx",
    "deeplearning",
    "computer-vision",
    "gpu-acceleration",
    "frame-interpolation",
    "super-resolution"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "numpy",
    "opencv-python",
    "tqdm",
    "onnx",
    "onnxruntime-directml",
    "pycuda",
    "nvidia-nvcomp-cu12",
    "brotli"
]

[project.urls]
"Homepage" = "https://github.com/zpeng11/ezvtuber-rt"
"Bug Tracker" = "https://github.com/zpeng11/ezvtuber-rt/issues"
"Source Code" = "https://github.com/zpeng11/ezvtuber-rt"
